chore: moving date formatting methods into its own file
