+++ pages/wards/visit-start.js
@@ -6,8 +6,7 @@ import retrieveVisitByCallId from "../../src/usecases/retrieveVisitByCallId";
 import propsWithContainer from "../../src/middleware/propsWithContainer";
 import { useState } from "react";
 import Error from "next/error";
-import formatDate from "../../src/helpers/formatDate";
-import formatTime from "../../src/helpers/formatTime";
+import { formatDate, formatTime } from "../../src/helpers/formatDatesAndTimes";
 import verifyToken from "../../src/usecases/verifyToken";
 import { WARD_STAFF } from "../../src/helpers/userTypes";
 import { v4 as uuidv4 } from "uuid";
