--- src/helpers/formatDate.js
@@ -1,4 +1,0 @@
-import moment from "moment";
-
-const formatDate = (time) => moment(time).format("D MMMM YYYY");
-export default formatDate;
