--- src/helpers/formatTime.js
@@ -1,6 +1,0 @@
-import moment from "moment";
-
-const formatTime = (time, timeFormat = "h:mma") =>
-  moment(time).format(timeFormat);
-
-export default formatTime;
